requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
line-bot-sdk==3.16.1
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.68.2
packaging==24.2
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
